refactor(artist-request): rename and reorganize artist request use case files
- Rename HasUserSubmittedRequestUseCase to CheckUserArtistRequestUseCase for clarity
- Update imports and references in related files
- Improve code formatting in container and controller files
