+++ services/user-admin-service/src/application/usecases/user/artist-request/CheckUserArtistRequestUseCase.ts
@@ -1,9 +1,9 @@
+import { ArtistRequest } from "@prisma/client";
 import { NotFoundError } from "art-chain-shared";
 import { USER_MESSAGES } from "../../../../constants/userMessages";
 import { IUserRepository } from "../../../../domain/repositories/user/IUserRepository";
 import { IArtistRequestRepository } from "../../../../domain/repositories/user/IArtistRequestRepository";
 import { ICheckUserArtistRequestUseCase } from "../../../../domain/usecases/user/artist-request/ICheckUserArtistRequestUseCase";
-import { ArtistRequest } from "@prisma/client";
 
 
 export class CheckUserArtistRequestUseCase implements ICheckUserArtistRequestUseCase {
