+++ services/user-admin-service/src/infrastructure/container/user/artistRequestContainer.ts
@@ -4,19 +4,24 @@ import { ArtistRequestRepositoryImpl } from "../../repositories/user/ArtistReque
 import { ArtistRequestController } from "../../../presentation/controllers/user/ArtistRequestController";
 
 import { CreateArtistRequestUseCase } from "../../../application/usecases/user/artist-request/CreateArtistRequestUseCase";
-import { CheckUserArtistRequestUseCase } from "../../../application/usecases/user/artist-request/HasUserSubmittedRequestUseCase";
+import { CheckUserArtistRequestUseCase } from "../../../application/usecases/user/artist-request/CheckUserArtistRequestUseCase";
 
 // Repositories
 const userRepo = new UserRepositoryImpl();
 const artistRequestRepo = new ArtistRequestRepositoryImpl();
 
 // Use Case
-const createArtistRequestUseCase = new CreateArtistRequestUseCase(userRepo, artistRequestRepo);
-const checkUserArtistRequestUseCase = new CheckUserArtistRequestUseCase(userRepo, artistRequestRepo)
+const createArtistRequestUseCase = new CreateArtistRequestUseCase(
+  userRepo,
+  artistRequestRepo
+);
+const checkUserArtistRequestUseCase = new CheckUserArtistRequestUseCase(
+  userRepo,
+  artistRequestRepo
+);
 
 // Controller
 export const artistRequestController = new ArtistRequestController(
   createArtistRequestUseCase,
   checkUserArtistRequestUseCase
-
 );
