+++ services/user-admin-service/src/presentation/controllers/user/ArtistRequestController.ts
@@ -1,13 +1,19 @@
 import { Request, Response, NextFunction } from "express";
-import { CreateArtistRequestUseCase } from "../../../application/usecases/user/artist-request/CreateArtistRequestUseCase";
 import { HttpStatus } from "art-chain-shared";
+
 import { ARTIST_MESSAGES } from "../../../constants/artistMessages";
-import { CreateArtistRequestDto } from "../../../domain/dtos/user/CreateArtistRequestDto";
-import { validateWithZod } from "../../../utils/zodValidator";
-import { createArtistRequestSchema } from "../../../application/validations/user/createArtistRequestSchema";
 import { USER_MESSAGES } from "../../../constants/userMessages";
-import { CheckUserArtistRequestUseCase } from "../../../application/usecases/user/artist-request/HasUserSubmittedRequestUseCase";
+
 import { IArtistRequestController } from "./interfaces/IArtistRequestController";
+
+import { CreateArtistRequestDto } from "../../../domain/dtos/user/CreateArtistRequestDto";
+
+import { validateWithZod } from "../../../utils/zodValidator";
+
+import { createArtistRequestSchema } from "../../../application/validations/user/createArtistRequestSchema";
+
+import { CreateArtistRequestUseCase } from "../../../application/usecases/user/artist-request/CreateArtistRequestUseCase";
+import { CheckUserArtistRequestUseCase } from "../../../application/usecases/user/artist-request/CheckUserArtistRequestUseCase";
 
 
 export class ArtistRequestController implements IArtistRequestController {
@@ -33,13 +39,15 @@ export class ArtistRequestController implements IArtistRequestController {
     try {
       const userId = req.headers["x-user-id"] as string;
       if (!userId) {
-        return res.status(HttpStatus.BAD_REQUEST).json({ message: USER_MESSAGES.USER_ID_REQUIRED })
-      };
+        return res
+          .status(HttpStatus.BAD_REQUEST)
+          .json({ message: USER_MESSAGES.USER_ID_REQUIRED });
+      }
       const result = validateWithZod(createArtistRequestSchema, req.body);
 
       const { bio, phone, country } = result;
 
-      const dto: CreateArtistRequestDto = {userId, bio, phone, country}
+      const dto: CreateArtistRequestDto = { userId, bio, phone, country };
 
       const request = await this._createArtistRequestUseCase.execute(dto);
 
@@ -69,17 +77,20 @@ export class ArtistRequestController implements IArtistRequestController {
     try {
       const userId = req.headers["x-user-id"] as string;
       if (!userId) {
-        return res.status(HttpStatus.BAD_REQUEST).json({ message: USER_MESSAGES.USER_ID_REQUIRED })
-      };
+        return res
+          .status(HttpStatus.BAD_REQUEST)
+          .json({ message: USER_MESSAGES.USER_ID_REQUIRED });
+      }
 
-       const { alreadySubmitted, latestRequest } = await this._checkUserArtistRequestUseCase.execute(userId);
+      const { alreadySubmitted, latestRequest } =
+        await this._checkUserArtistRequestUseCase.execute(userId);
 
       return res.status(HttpStatus.OK).json({
         message: ARTIST_MESSAGES.REQUEST_SUBMITTED_SUCCESS,
         data: {
           alreadySubmitted,
-          latestRequest
-        }
+          latestRequest,
+        },
       });
     } catch (error) {
       next(error);
